+++ core/src/main/java/arez/ArezTestUtil.java
@@ -63,7 +63,8 @@ public final class ArezTestUtil
     ( (ArezLogger.ProxyLogger) ArezLogger.getLogger() ).setLogger( null );
     Transaction.setTransaction( null );
     Transaction.resetSuspendedFlag();
-    Arez.reset();
+    ArezZoneHolder.reset();
+    ArezContextHolder.reset();
   }
 
   /**